requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
future==0.18.2
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
grpcio==1.26.0
h5py==2.10.0
idna==2.8
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
logger==1.4
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
Pillow==7.0.0
promise==2.3
protobuf==3.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.13.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-datasets==3.0.0
tensorflow-estimator==2.1.0
tensorflow-metadata==0.21.2
tensorflow-probability==0.8.0
termcolor==1.1.0
tqdm==4.45.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
